Switch to u32 for chances

+++ src/lib.rs
@@ -50,11 +50,11 @@ pub fn roll_walker_table(quantity: i32, index_weights: Float32Array) -> String {
 }
 
 #[napi]
-pub fn simulate_chances(rolls: u32, chance: u16) -> u32 {
+pub fn simulate_chances(rolls: u32, chance: u32) -> u32 {
   let mut success_count = 0;
   let mut rng = fastrand::Rng::new();
   for _ in 0..rolls {
-    if rng.u16(0..chance) == 0 {
+    if rng.u32(0..chance) == 0 {
       success_count += 1;
     }
   }
